Remove trace.[trace_id/span_id] from transaction contexts (#2095)
# Context

We currently store trace_id and span_id in two places on the transactions table:

* As columns
* As keys and values on the contexts table

This storage is redundant and also non-compressible since these ids are randomly generated.
Solution

This PR makes it so the message processor that writes to the table drops it from the contexts table. It will not be queried due to the changes in #2074

# Test Plan

Added to the test_transaction_processor test that validates trace_id and span_id being promoted to the column but not being stored in the contexts
    
# Rollout Concerns

If we start doing this, we have to make sure that we are no longer querying the contexts[trace.trace_id] column. This was verified in the querylog with the following query returning no rows:
```
SELECT COUNT (*) FROM (
	SELECT
	    request_id,
	    referrer,
	    clickhouse_queries.sql
	FROM
		querylog_local
	ARRAY JOIN
		clickhouse_queries
	WHERE
	    (
	        like(clickhouse_queries.sql, '%indexOf(contexts.key, \'trace.span_id\')%') OR
	        like(clickhouse_queries.sql, '%indexOf(contexts.key, \'trace.trace_id\')%') 
	
	    ) AND
	    timestamp >= toDateTime('2021-09-07 00:00:00', 'America/Los_Angeles')
) AS T
```
This query does return rows for the previous week (before #2074 was deployed)