feat(web): Entity-specific insertion endpoint for tests (#2101)
The existing API endpoint /tests/<dataset:dataset>/insert allows writing
data to the default entity of each dataset. In the case of metrics, this
means that only metrics_sets can be written.

This PR introduces an additional endpoint
/tests/entities/<entity:entity>/insert, which allows the client to
define the target entity explicitly.

This is required to run tests in getsentry/sentry#28598.