feat: Add original message timestamp to commit log headers - take 2 (#2106)
This is the same as the original version in https://github.com/getsentry/snuba/pull/2093 except it uses arroyo 0.0.4 which handles the case where the `orig_message_ts` header is missing when decoding messages. Previously, old messages still in the pipeline that were produced with an older version of arroyo would be missing the header which caused the consumer to crash. Once there are no more old messages in our pipeline we can enforce the presence of the header again.