fix: Fix queries on columns with different underlying types
Queries involving columns with different internal and external
types did not behave properly when it included a condition involving
one of the >, >=, <. <=, LIKE or NOT LIKE operators.

The BaseTypeConverter contained an optimization that attempted to use
the internal column for the comparison however this is not correct
as the comparison returns completely different results depending on
the type.

This change drops the optimizations on all comparisons except =, !=,
IS NULL and IS NOT NULL. Queries containing a mix of optimizable and
non optimizable conditions are treated as unoptimized and the
external type is used.