SnubaException: A new exception class for all your exception needs (#2109)
### Context

Following up from #2102 this PR implements a serializable exception class which can be serialized and reraised upon deserialization

### Solution

SerializableException is an extension of the Exception class that intends to be the base exceptions for all exceptions in Snuba. All Exception subclasses have already been replaced in this PR. In addition to serializability, SerializableException also adds whether the error should be reported to sentry as a first class citizen.