Communicate read-through cache exceptions to waiting clients (#2102)
### Problem
We have a lot of ExecutionErrors happening for reasons that we don't really know. This PR reports those errors back to clients waiting for results of duplicate queries

### Proposed Solution

If an exception happens serialize the exception using SerializableException and put that in redis. Deserialize it on the other side and re-raise it

While I was here,  I made the redis lua script more readable as well