feat(replacements) Add a simple rate limiter to cap the amount of INSERT queries to clickhouse (#2122)
The replacer, today, does not cap in any way the amount of INSERT queries sent to clickhouse. This was not too dangerous before re-processing was built as replacement operations were scattered and rare.
Now we can have a constant amount of replacements being issued for reprocessing over a long period of time. These were used to be batched to reduce the impact on clickhouse but now the batch size has been reduced to only 10 and we can have large amount of replacements issued at high rate over a long period of time.
This can actually cause clickhouse to accumulate inactive parts and to be unable to keep up with merge. Which would cause an outage.
This adds a simple safe guard by capping the amount of INSERT per second you can run on any replacer so that if the issue happens, in the worst case scenario, we would backlog on Kafka, which is still a lot better than a Clickhouse outage.

In the meantime the ingest team is working on re-introducing the batching.