Cache error values in read through cache for only 100ms (#2125)
### Context

By storing the error values in the cache when errors happen, we end up caching transient errors for way too long. In order to temporarily fix this, error values will have a TTL of only 1s instead of the full 7 minutes that normal result queries have. This will be fixed more properly very shortly.

### Solution

* Make the expiry time for errors values 1s

### Test plan
Add a test that assures that errors values are quickly flushed from the cache. Even when we fix this properly that test should still behave the same way