feat(subscriptions): Subscription data store uses entity rather than dataset name (#2121)
The RedisSubscriptionDataStore now stores subscriptions by entity name instead
of dataset name. This is a step towards making all subscriptions based on
an entity rather than a dataset to align with our API. Since multiple
entities (such as events and transactions) are likely to eventually become
part of the same dataset, we should ensure that subscriptions are stored separately
for each entity even if they later become part of the same dataset.

Since all dataset and entity names for our current subscriptions are the same,
no migration is required.