Communicate read through cache errors in separate key  (#2128)
### Context
 
The read through cache was changed to communicate errors to waiting clients by storing the error value in place of the result. The result however has a TTL of 7 minutes in production meaning that if a query hits an error, the same error will be propagated for the next 7 minutes.

### Solution

Store error values in a key of their own. If a waiting process receives a notification and there is nothing in the value key, it checks the error key. The value key only holds true result values.

### Test plan

There is already a test for transient errors not being re-raised introduced in #2125 but now the time between tasks can be much smaller as we're not replying on the TTL to handle that invalidation logic for us.