feat(subscriptions): Add tick buffer processing strategy
Adds a processing strategy that buffers ticks before submitting them
to the next processing step in the pipeline. This is not currently
being used anywhere but will be used by the future subscription scheduler.

The tick buffer has two modes, which correspond to the two different
ways we will be doing subscription scheduling.

- Immediate mode immediately submits a tick to the next step when
we consume it. The use case for this is for subscriptions on topics
that are semantially partitioned by project ID. When we get a tick,
we can safely schedule it immediately since all the project's messages
are in the same partition and order is already guaranteed within a
single project.

- Wait-for-slowest-partition mode buffers ticks in memory and only
submits a tick to the next step once the upper timestamp of that tick
has been reached on every partition. The use case for this mode is for
subscriptions on topics that are partitioned randomly or by something other
than the project ID. Waiting for all partitions to reach a timestamp provides
a guarantee that all data up to that point has been written to storage.
This can also be used to run queries spanning multiple projects, as waiting
ensures our data up to that timestamp is complete before running the
subscription query. An integer value for `max_ticks_buffered_per_partition`
must be provided for this strategy. Since we cannot infinitely buffer ticks,
once this max value is reached, ticks will be submitted to the next step even
if not all partitions have reached that point yet.

The TickBuffer strategy currently records the lag between the slowest and
fastest partition (similarly to what is done in the temporary `MeasurePartitionLag`
strategy). In future, we will likely want to record more metrics here, such
as the count of buffered ticks. This strategy will likely also need to
be responsible for committing offsets.