feat(subscriptions): Use scheduler command to debug commit log issues (temporarily)
Upon turning on the scheduler process in production for testing, we experienced
a lot of issues with invalid intervals. The lower bound of the interval constructed
by the CommitLogTickConsumer was frequently higher than the upper bound, which is invalid.

One potential cause of this could be messages being out of their expected order in
the commit log. Though this is known to be possible due to consumer crashes, rebalancing, etc
the frequency with whicih this appeared to happen seemed unusual.

This PR aims to change the scheduler consumer to simply collect more information about
commit log ordering to understand how common out of order messages might be.
For each message consumed, we simply record whether it is earlier or later than the
last one we received for that partition.

In order to avoid introducing other bugs or complexities to interval calculations,
it tries to be as minimalistic as possible and uses just a standard consumer, not the
tick consumer.