fix(subscriptions): Tick consumer filters messages by consumer group (#2139)
Previously the CommitLogTickConsumer was attempting to build ticks from all of the
consumer groups that are producing to the commit log. Since we have two consumer groups
on the `snuba-commit-log` (events and transactions) this resulted in an attempt to build
a lot of incorrect ticks, and caused a lot of InvalidRangeError. The tick consumer should
only follow the messages of one consumer group.

This PR also adds the --followed-consumer-group option to the subscription-scheduler
command. It is temporarily set to not required yet so this can be rolled out without causing
the currently running scheduler to crash. After we have have ensured we are always passing
this argument and our ticks are always properly ordered, we can switch the scheduler back
to actually use the tick consumer.