feat(subscriptions): Add scheduler mode to stream loader
The Kafka stream loader defines the scheduler mode that applies to
the subscriptions on a specific storage. The scheduler mode is
associated in code with the storage / topic and should not need
to be changed by users when running subscriptions. Generally speaking,
immediate mode will apply to semantically partitioned topics and
wait-for-slowest mode applies for non semantically partitioned topics
where we will wait for all partitions to reach a timestamp before
executing the subscriptions associated with that particular timestamp.
The mode must be specified if subscriptions are supported for that storage.