feat(subscriptions): Add scheduler mode to stream loader (#2140)
The Kafka stream loader defines the scheduler watermark mode that applies to
the subscriptions on a specific storage. The scheduler mode is
associated in code with the storage / topic and should not need
to be changed by users when running subscriptions. Generally speaking,
"partition" mode will apply to semantically partitioned topics and
"global" mode applies for non semantically partitioned topics
where we will wait for all partitions to reach a timestamp before
executing the subscriptions associated with that particular timestamp.
The mode must be specified if subscriptions are supported for that storage.