perf(subscriptions): Add minimum tick interval option for subscriptions
This is an experiment to see if we can improve the performance of the subscription
worker by creating fewer ticks which span larger intervals, which will result in
less calls to later parts of the subscriptions pipeline and messages submitted
to the processing strategy.

The subscriptions CLI now support a --min-tick-interval-ms option, which can
be a value between 1 and 1000. If a value is passed, a tick will only be
created when we get messages spanning an interval larger than this range. This
results in fewer tick message but each covering a larger range.

We should only attempt to use this option (at least for now) on the experimental
sessions subscriptions consumer, and not events or transactions.