feat(metrics) Expose an abstract aggregation functions in the metrics entities (#2142)
This solves two problems:

The previous implementation where a virtual column was hiding the aggregation state merge function did not allow for conditional aggregations like countIf or avgIf, which for aggregated tables have to be expressed as countMergeIf and avgMergeIf

Having the aggregation hidden behind a virtual column confused several engineers (who were adding a group by bu it did not seem the query included an aggregation.

We do not want to expose the specific aggregation function as we may want to experiment with different data sketches for performance. The product should not care about them and we should also not expose the implementation details of the aggregation state merges (the -Merge keyword).
So this exposes the aggregation function but they are abstract and translated into the concrete ones by the translators. This also allows for the conditional aggregations.