feat(subscriptions): Add consumer queue options to subscriptions
This adds the ability to override the default queued.min.messages and
queued.max.messages.kbytes librdkafka settings on the subscriptions consumer,
allowing us to provide different values for each subscriptions consumer.

The immediate motivation for this change is to try and reduce the consumer group
lag on the sessions consumer by increasing the buffer size.

We already support these two options in the main snuba consumer and replacer
processes, this change adds the same settings to subscriptions.