feat(subscriptions): Use tick buffer in scheduler, add support for committing offsets
This change updates the TickBuffer processing strategy to commit offsets.
The strategy takes a commit function in the contructor and calls it only
once every partition has reached that offset. This behavior is the same
regardless of whether the scheduler is in 'global' or 'partition' mode.
Whilst scheduling can happen immediately or on some delay depending on
the mode and buffer configurations, commiting offsets always happens
only when the slowest partition has reached a particular offset. This
guarantees that every subscription is scheduled at least once even in the
event of a scheduler restart.

The subscription scheduler has also been updated to use the new tick
buffer processing strategy. The mode is temporarily hardcoded to
'global' just for testing and experimentation.
The correct value should be provided by the stream loader.