feat(suspect-spans): Add array join optimizations on spans queries (#2117)
Similar to the array join optimization on tags, this change rewrites array join
queries on spans as a single array join with the optional array filter
optimization. It also adds the necessary conditions in order to take advantage
of the bloom filter index available on the spans.op and spans.group columns.