build(feat): Support aarch64 image building
In order to support building a Docker image that can run on `arm64/aarch64` we need to install or build Python binary wheels.

`sentry-relay` version `0.8.9` has Python wheels for `aarch64`, thus, we need to bump the version.

In order to build confluent-kafka from source, we need to add `librdkafka-dev` since there's no `aarch64` wheel for it.
In order for the code to find the library, we need to leave it on disk, otherwise, the image won't build.

> #13 2.834   File "/usr/local/lib/python3.8/site-packages/confluent_kafka/deserializing_consumer.py", line 19, in <module>
> #13 2.834     from confluent_kafka.cimpl import Consumer as _ConsumerImpl
> #13 2.834 ImportError: librdkafka.so.1: cannot open shared object file: No such file or directory