feat(subscriptions): Scheduled task encoder should encode a KafkaPayload (#2152)
The scheduled task encoder is now responsible for encoding a ScheduledTask
into a KafkaPayload and vice versa. This was it can be used more easily
by the subscriptions scheduler to encode a scheduled task so it can be
passed directly to a Kafka producer, and by the subscriptions executor
to decode a message received from the consumer directly back into a ScheduledTask.

This codec is not currently being used anywhere but will be integrated into
the scheduler (and eventually the executor as well) in a follow up.