feat(snql) Remove DelegateSubscription and LegacySubscription
The Delegate class was added as a way to easily test moving from Legacy
subscriptions to SnQL subscriptions. With it we had subscriptions with both
Legacy and SnQL query parameters and so using a setting could toggle back and
forth between the two formats.

Now that all subscriptions are SnQL subscriptions (thanks to the migration
https://github.com/getsentry/sentry/blob/master/src/sentry/migrations/0237_recreate_subscriptions_in_snuba.py)
we can safely deprecate all this code, since every subscription will have a SnQL
query in it.

Since we no longer need to toggle back and forth between query formats, we can
remove the entire delegate class. Technically the subscriptions in Redis will
have legacy query fields in them, but they will be ignored. It doesn't seem
worth another full migration just to remove those fields.