feat(subscriptions): Add a temporary scheduler that mimics higher load (#2171)
The LoadTestingSubscriptionScheduler is similar to the existing SubscriptionScheduler
class but it simulates a higher load by multiplying the number of actual subscriptions
by some `load_factor`. It is not being used anywhere yet but will be
integrated into the scheduler process in a following steps.

Also adds the --load-factor option to the subscription CLI.

This is just for testing scheduler capacity and should be completely removed
before the scheduler is used for any real purpose.