perf(consumers): Route messages based on kafka headers
The change introduces a new pre filter which looks at kafka headers
and decides whether the message should be dropped or not.

It is used with the errors and transactions consumers which have
messages on the same topic. It should improve their performance
since they don't need to deserialize the message anymore to
determine whether to drop message or not.