feat(subscriptions): Add processing strategy to schedule subscriptions
Adds a processing strategy that takes a `CommittableTick` - that is - a tick
message with a boolean value indicating whether its offset can be safely
committed or not, and does two things:
- schedule all of the subscriptions that need to be run given the tick's interval and partition
- commit the offset if the `CommittableTick` specifies should_commit = True

This strategy is supposed to come directly after the TickBuffer and ProvideCommitStrategy
strategies and actually schedule the subscriptions.
https://github.com/getsentry/snuba/pull/2151 contains the detail about how the
`CommittableTick` is produced from a `Tick`

This code is not used yet and will be integrated into the scheduler as a follow up
step after https://github.com/getsentry/snuba/pull/2151.