feat(config): check value types
The current config system does not check the type of the values being set
for keys. This allows for user mistakes where the type of the value could
be changed unknowingly. Added type checking for values and an option to
override it in case the existing value type was incorrectly set.