ref(legacy) Remove legacy query endpoint (#2173)
The legacy query endpoint is no longer being used. Remove it, and make the SnQL
parser the default. This also removes the concept of the query Language, since
now there is only one API language. This also slightly restructures the
build_request by pushing custom processing up a level, which cleans up the code
a bit.