perf(consumer) Add a mock version of the consumers for load test (#2179)
This introduces a mock batch writer and a mock replacer writer to allow us to run load tests of our consumers on production data in an idempotent way without writing anything on Clickhouse.

The idea is to create a dedicated consumer cluster.
Let the consumer group backlog for some time
Then start the consumer and measure how long it takes to the consumer to burn the backlog.
This can be done to evaluate the current headroom (if we do not burn the backlog we have a problem) as well as to compare parameters (processes, batch size, replicas count)

The mock batch writer adds a configurable delay to simulate the clickhouse write delay (which we already measure). We do not do the same for replacement as they are so rare compared to batch writes that the replacement production delay would not impact the test.