ref(legacy) Remove parse_query function
Remove the `parse_query` function since it is no longer used. Most of the work
in this PR was in refactoring all the tests that used it to test various parts
of the AST parsing. In some cases the queries were rewritten to be SnQL, in
others the snuba-sdk was used to convert the legacy to SnQL. There is now an
open tech debt story to convert all these tests to be pure SnQL.