feat(subscriptions): Use tick buffer in scheduler, properly commit offsets (#2151)
This change adds a new ProvideCommitStrategy processing step that is
responsible for determining whether the offset of a message can be
safely committed or not. It is designed to be chained together with the tick
buffer strategy - it get passed a `Tick` and submits that message to the next
step as a `CommittableTick` - basically that same tick with a True or False
`should_commit` value indicating whether it can be safely committed or not.

In order to guarantee that all subscriptions are executed at least once, we can
only safely commit offsets based on the slowest partition. This ensures that even
if the scheduler is restarted and loses it's state we do not skip any intervals
on any partitions.

The subscription scheduler has also been updated to use the new tick
buffer processing strategy together with the commit strategy. The mode
is temporarily hardcoded to 'global' just for testing and experimentation.
The correct mode value should be provided by the stream loader.
The buffer size can be configured via the 
`SUBSCRIPTIONS_DEFAULT_BUFFER_SIZE` and
`SUBSCRIPTIONS_ENTITY_BUFFER_SIZE` settings. The default is 10000.

We are monitoring this in production via the `partition_lag_ms` and consumer
group lag metrics.