To be removed: query code formatter (#2248)
It can be very tedious to write out queries by hand so you can test them this PR does some amount of the legwork to make that a reality. I am merging this branch into the feature I'm working on but will remove it before a final review.

The current code is not copy-pastable as-is and also uses a hack for formatting (imports black and calls it directly). This is just to unblock me in the specific feature I am doing. This code will be removed later (but the PR will remain for future reference for those who want to see what I did)