feat(subscriptions): Also produce tick to the scheduled subscriptions topic
We need to also produce the tick along with the scheduled subscription data
to the scheduled subscriptions topic. This is required by the executor
in order to actually build the request and execute the query. It cannot be
omitted as many subscriptions will later modify the query to add an offset
condition during execution. We are likely to also use it to run comparisons
between the new and existing pipeline during roll out.

An alternative would be to have the scheduler build the entire request
and the executor only run it, but I did not go down this route as it:
- puts more work on the scheduler (which is harder to scale than executors)
- added complexity of serializing and deserializing the entire request object
(in comparison scheduled task + tick are much simple structures)

Since we don't have an executor that is decoding any of the messages yet,
this change should be able to safely merge and deploy.