feat(suspect-spans): Change exclusive time to Float32 (#2231)
Using Float32 is enough precision for the exclusive time column. Dropping the
original column and starting over. This column is only internal at the moment so
it is safe to drop any existing data. Before dropping the column, make sure the
`input_format_skip_unknown_fields` option is enabled and/or the processors are
not emitting the span columns to avoid errors in the transaction processors.
Additionally, make sure that `materialize_ttl_after_modify` is disabled prior
modifying the TTL to avoid massive merges.