feat(suspect-spans): Stop writing to old exclusive time column
With the new exclusive time column being written to, there is no need to keep
writing to the old exclusive time column. This emits all 0s to it to save on
space. It is not possible to stop emitting this value entirely until the column
is dropped due to ClickHouse expecting an array of equal length for all columns
in the nested column.