feat(subscriptions): Write entity to scheduled topic
In order for the subscriptions executor to be able to handle more than one
entity on a scheduled topic we need to include the entity information into the
Kafka message. Currently the scheduled subscription task codec depends on
knowing the entity key in order to decode the message - this will no longer be
possible when more than one entity's subscriptions are in the same topic.

Since the executor is not running yet and we are never decoding messages
we can safely make this change.