feat(subscriptions): Write entity to scheduled topic (#2257)
In order for the subscriptions executor to be able to handle more than one
entity on a scheduled topic we need to include the entity information into the
Kafka message. Currently the scheduled subscription task codec depends on
knowing the entity key in order to decode the message - this information will no
longer be available when more than one entity's subscriptions are in the same topic
(and will be passed via the executor CLI command).

Since the executor is not running yet and we are never decoding messages
we can safely make this change.