feat: Add subscriptions executor command (#2255)
Subscription execution forms the second part of the subscriptions pipeline.
The executor's job is to consume the scheduled subscriptions from the scheduled
subscription topic, run the queries on ClickHouse and produce the results to
the results topic for Sentry.

This change simply introduces the executor CLI command. It builds the consumer
and processing strategy, however that strategy doesn't really do anything yet
except commit offsets. The processing strategies that are responsible for actually
running the query on ClickHouse and producing the results will be added separately.

Currently only a single dataset/entity pair is supported by the subscriptions
executor and a separate executor needs to be run for each entity.