feat(subscriptions): Executor command takes multiple entities
The executor command can now take multiple entities as long as those entities:
- match the dataset value which is passed
- have the same scheduled topic defined
- have the same result topic defined

In the future we may want to make this more flexible by providing a way to override
the scheduled and result topics - but for now they must all be the same in the code.

The entity name is not being used yet by the executor process yet. Later we will filter
scheduled subscriptions with entity names that do not match any of the values provided.