feat(perf) Optimize away redundant tag conditions (#2256)
### Context
We occasionally fail to apply the tags hashmap optimization because there is a redundant clause in there. 

The problem boils down to this:
```
        valueOf('my_tag') != '' AND valueOf('my_tag') == "something"
                          ^                            ^
                          |                            |
                      existence check               value check
```

The tags hash map is stored with elements `k=v` ([here](https://github.com/getsentry/snuba/blob/master/snuba/datasets/storages/tags_hash_map.py#L7-L11)) meaning that checking `exists("my_tag")` is not possible with the hashmap as it exists today. 

The mapping optimizer will not apply the hashmap optimization if the `tags` column has to be loaded anyway. This is because the cost of scanning an extra column in clickhouse is higher than the speedup gained from using the optimization

### Solution

Within the mapping optimizer, write a function that removes the redundant conditions first, then apply the mapping optimzer as before

### Test plan

* Unit test coverage of new optimizer functionality
* end to end query pipeline test on problematic snql query to see that the optimization is applied
* Experiment configured (to be verified in discover)