perf: Omit redundant information from scheduled subscription message
This change attempts to improve the performance of the scheduler by
reducing the size of the scheduled message and minimizing the amount
of work that the encoder needs to do. Previously we were encoding
and producing the entire tick to the scheduled subscription topic.
However the only part we actually need in order to build the request
is the upper offset.