feat(subscriptions): Execute query submits SubscriptionTaskResult to next step
The next processing step after ExecuteQuery will be responsible for encoding the
SubscriptionTaskResult and producing it to the result topic. Since we are going to
reuse the SubscriptionTaskResultEncoder (from the existing pipeline) we need to use
the same SubscriptionTaskResult class (previously they were two distinct classes).

Since the future is already completed by the time we pass the task to the next step
we should also just submit the actual SubscriptionTaskResult instead of the
SubscriptionTaskResultFuture.