feat: Sample queries run by subscriptions executor (#2275)
The subscriptions executor samples queries according to the rate
provided by `executor_sample_rate`. This is to avoid overloading
ClickHouse with queries while we are running both the old and new
pipelines in parallel. Sampling is based on a hash of the subscription
ID so that the same subscriptions are always being executed. The default
is 0.0, i.e. no queries are passed through the executor pipeline or run
on ClickHouse.