feat(subscriptions): Executor uses the ExecuteQuery strategy
Integrates the ExecuteQuery strategy in the subscriptions executor.
Note that currently the `executor_sample_rate` option is not set
so all scheduled subscriptions will be filtered at the start of the
pipeline, and even their offsets will not be committed.

This is to avoid running duplicate queries on ClickHouse until we
are ready to begin testing.