feat(admin): Add UI component that loads the query node data
Adds UI component which makes an API call for the storage/node data from
the `/clickhouse_nodes` endpoint. The storage names and ClickHouse node
list are populated into 2 dropdowns. We are not yet doing anything with
the user's selections except keeping track of them in the component's
state.

The API now alphabetizes node data by storage so it is more readable.