feat(subscriptions): Add the strategy to produce a subscription result
This change adds the ProduceResult strategy. It's job is to take a
SubscriptionTaskResult, encode the message to a Kafka payload and produce
it to the provided result topic. Once producing is done, offsets are
committed. This strategy is designed to be chained together with the
ExecuteQuery strategy (which provides the SubscriptionTaskResult) in the
subscriptions executor.

The approach taken is very similar the ProducedScheduledSubscriptionMessage
strategy in the subscription scheduler which encodes and produces the
messages for the scheduled topic.

This change only adds the strategy - it is not being used anywhere yet
and will be integrated into the executor as a follow up.