feat(tracing) Change the return type of the native execute  (#2279)
It is very difficult to capture new data from Clickhouse with the current return
value from the function. The current wrapper around the clickhouse-driver
library returns a simple list of results, that gets overloaded with column types
when that meta information is returned. Change the return type to be an object,
so that any data that we might want to propogate up from Clickhouse becomes easy
to add without having to go through this process again.